Make some parameters slower to change
Make some parameters (brightness, animation speed, and animation spread)
slower to change when changing them with the rotary encoder. This makes
them a bit easier to adjust.

Also move brightness to the front as it is very clear what it does
(makes it hopefully easier for people to discover what the button is
for).